refactor: Move form input functions into a separate components file.
